Moves payment-step scenarios into dedicated file
